Move status (filter) text so it's with Toolbar
