+++ app/src/main/res/layout/activity_drawer.xml
@@ -11,7 +11,7 @@
         android:layout_height="match_parent"
         android:orientation="vertical">
 
-        <include layout="@layout/toolbar_actionbar" />
+        <include layout="@layout/toolbar_directory" />
 
         <FrameLayout
             android:layout_width="match_parent"
