+++ app/src/main/res/layout/fragment_recyclerview.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
@@ -8,21 +9,6 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:visibility="gone">
-
-        <TextView
-            android:id="@+id/status"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:textColor="@android:color/white"
-            android:textSize="17sp"
-            android:gravity="center_vertical|start"
-            android:paddingBottom="12dp"
-            android:paddingLeft="72dp"
-            android:paddingStart="72dp"
-            android:paddingRight="16dp"
-            android:paddingEnd="16dp"
-            android:visibility="gone"
-            android:fontFamily="sans-serif-light" />
 
         <android.support.v7.widget.RecyclerView
             android:id="@android:id/list"
@@ -46,7 +32,8 @@
                 android:layout_gravity="center_horizontal"
                 android:src="?empty_image"
                 android:scaleType="fitXY"
-                android:layout_marginBottom="16dp" />
+                android:layout_marginBottom="16dp"
+                tools:ignore="ContentDescription" />
 
             <TextView
                 android:id="@+id/emptyText"
