+++ app/src/main/res/layout/toolbar_actionbar.xml
@@ -3,7 +3,7 @@
     android:id="@+id/toolbar_actionbar"
     android:layout_width="match_parent"
     android:layout_height="?actionBarSize"
-    android:background="?colorPrimary"
+    android:elevation="4dp"
     app:contentInsetStart="72dp"
     app:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar"
     app:popupTheme="@style/ThemeOverlay.AppCompat.Light"
